Added AssignmentTable class that represents the assignments graph. Changes in occlusion handling. Added new constructors for Track.
