+++ Developing/Tracker.cpp
@@ -2,220 +2,7 @@
 // Created by Ars Poyezzhayev on 06.12.16.
 //
 #include "Tracker.h"
-
-class Assignments {
-    /**
-     * It's optimized analogue of the matrix of assignments, e.g.
-     *
-     *          Tracks
-     *        1   2   3   4   5   6   7
-     *      ----------------------------
-     * D  1|  *   A   *   *   *   *   * |
-     * e  2|  *   *   *   *   *   *   * | (U)
-     * t  3|  *   *   *   S   *   *   * |
-     * e  4|  *   *   MS  *   M   *   * |
-     * c  5|  M   *   *   *   *   M   * |
-     * t  6|  *   *   *   S   *   *   * |
-     * s  7|  *   *   S   *   *   *   * |
-     *      ----------------------------
-     *                                (U)
-     *                                (R)
-     * Where:
-     * A - assigned track to detection (one-to-one)
-     * U - unassigned track or detection (full line of *-s)
-     * M - merged tracks (n tracks per one detection)
-     * S - splitted track (n detections per one track)
-     * MS - merge and split (track that is merged with other n tracks and simultaniously is splitted to m tracks)
-     * R - track to remove
-     * (*) - no assignment
-     *
-     * It's represented as tracksToDetections and detectionsToTracks:
-     *                      1  2  3       4       5  6  7
-     * tracksToDetections: {5, 1, S(4,7), S(3,6), 4, 5, U}
-     *
-     *                      1  2  3  4       5       6  7
-     * detectionsToTracks: {2, U, 4, M(3,5), M(1,6), 4, 3}
-     *
-     * If tracks are splitted we should display it in the tracksToDetections list and if tracks are merged we consider it
-     * in the detectionsToTracks list.
-     */
-
-    /**
-     * Class that has a reference to the related assignment
-     */
-    class Assignment {
-    public:
-        int type;
-        bool isConsidered;
-        std::vector<int> additionalAssignments;
-
-        Assignment(int type) : type(type) {}
-
-    };
-
-public:
-    track_t distanceThreshold;
-    std::vector<Assignment> tracksToDetections;
-    std::vector<Assignment> detectionsToTracks;
-
-    // special types of assignments. The Assigned type is not specific, it's just a number of an assignment
-    enum type {
-        Unassigned = -1,
-        Merged = -2,
-        Splitted = -3,
-        ToRemove = -4
-    };
-
-
-public:
-    /**
-     * Transformation vectors maps elements of a matrix that contains only specific rows and columns of full matrix
-     * to the corresponding elements of the full matrix.
-     *
-     * Full:
-     *    1  2  3  4
-     *   ------------
-     * 1| A  B  C  D
-     * 2| E  F  G  H
-     * 3| J  K  L  M
-     * 4| N  O  P  Q
-     *
-     * Partial:
-     *       1(1)  2(3)  3(4)
-     *      ----------------
-     * 1(2)|  E     G     H
-     * 2(4)|  N     P     Q
-     *
-     * Transformation vectors:
-     * actualRows: [1] = 2, [2] = 4
-     * actualCols: [1] = 1, [2] = 3, [3] = 4
-     */
-
-    Assignments(size_t tracksSize, size_t detectionsSize, track_t distanceThreshold) :
-            distanceThreshold(distanceThreshold) {
-        for (int i = 0; i < tracksSize; ++i) {
-            tracksToDetections.push_back(Assignment(Unassigned));
-        }
-
-        for (int i = 0; i < detectionsSize; ++i) {
-            detectionsToTracks.push_back(Assignment(Unassigned));
-        }
-    }
-
-    void solve(distMatrix_t &costMatrix, int N, int M) {
-        std::vector<int> actualRows; //positions of unassigned detections
-        std::vector<int> actualCols; //positions of unassigned tracks
-
-        //occlusion handling works with detectionsToTracks and tracksToDetections
-
-
-        // Form transformation vectors
-        for (int i = 0; i < detectionsToTracks.size(); ++i) {
-            if (detectionsToTracks[i].type == Unassigned) {
-                actualRows.push_back(i);
-            }
-        }
-
-        for (int i = 0; i < tracksToDetections.size(); ++i) {
-            if (tracksToDetections[i].type == Unassigned) {
-                actualCols.push_back(i);
-            }
-        }
-
-        // Build cost matrix (costAdj) only for unassigned tracks and detections
-        size_t MAdj = actualRows.size();
-        size_t NAdj = actualCols.size();
-        distMatrix_t costAdj(NAdj * MAdj);
-
-        for (int i = 0; i < MAdj; ++i) {
-            for (int j = 0; j < NAdj; ++j) {
-                costAdj[i * NAdj + j] = costMatrix[actualRows[i] * N + actualCols[j]];
-            }
-        }
-
-        assignments_t assignmentVector;
-
-        // Solving assignment problem (tracks and predictions of Kalman filter)
-        DataAssociationAlgorithm HungarianAlg(Unassigned);
-        HungarianAlg.solve(costAdj, NAdj, MAdj, assignmentVector, DataAssociationAlgorithm::many_forbidden_assignments);
-
-        // -----------------------------------
-        // Remove assignment from pairs with large distance
-        // -----------------------------------
-        for (size_t i = 0; i < assignmentVector.size(); i++) {
-            if (assignmentVector[i] != Unassigned) {
-                if (costAdj[i*NAdj + assignmentVector[i]] > distanceThreshold) {
-                    assignmentVector[i] = Unassigned;
-//                    tracksToDetections[actualCols[i]]->skipped_frames = 1;
-                }
-//            } else {
-//                // If track have no assigned detect, then increment skipped frames counter.
-//                assignmentVector[i] = Unassigned;
-////                tracks[actualCols[i]]->skipped_frames++;
-            }
-        }
-
-        // Data association algorithm produces one-to-one correspondence between tracks and detections. We need to incorporate these results into the assignments
-        // After this step we have a full assignment matrix
-
-        for (int i = 0; i < assignmentVector.size(); ++i) {
-            if (assignmentVector[i] != Unassigned) {
-                assignTrackToDetection(actualCols[i], actualRows[assignmentVector[i]]);
-            }
-        }
-
-        /**
-         * I.e.: 7 tracks and 7 detections:
-         *                      1  2  3       4       5  6  7
-         * tracksToDetections: {5, 1, S(4,7), S(3,6), 4, 5, U}
-         *
-         *                      1  2  3  4       5       6  7
-         * detectionsToTracks: {2, U, 4, M(3,5), M(1,6), 4, 3}
-         */
-
-    }
-
-    void assignTrackToDetection(int track, int detection) {
-        tracksToDetections[track].type = detection;
-        detectionsToTracks[detection].type = track;
-    }
-
-    void merge(std::vector<int> &tracks, int detection) {
-        assert(tracks.size() > 1);
-        //TODO: do we need the sorting for the tracks??
-
-        for (int n : tracks) {
-            tracksToDetections[n].type = detection;
-            detectionsToTracks[detection].additionalAssignments.push_back(n);
-        }
-        detectionsToTracks[detection].type = Merged;
-    }
-
-    void split(std::vector<int> &detections, int track) {
-        assert(detections.size() > 1);
-        //TODO: do we need the sorting for the detections??
-        for (int n : detections) {
-            detectionsToTracks[n].type = track;
-            tracksToDetections[track].additionalAssignments.push_back(n);
-        }
-        tracksToDetections[track].type = Splitted;
-    }
-
-    // Returns positions of the detections and tracks which are not in the occlusions in the initial Cost matrix
-    //TODO: check the order of the detections and tracks. It could work only if the order is the same as in the initial matrix/
-    void getOneToOneDetections(std::vector<int> &detectionsPositions, std::vector<int> &tracksPositions) {
-        for (int i = 0; i < detectionsToTracks.size(); ++i) {
-            if (detectionsToTracks[i].type == Unassigned) {
-                detectionsPositions.push_back(i);
-            }
-        }
-        for (int i = 0; i < tracksToDetections.size(); ++i) {
-            if (tracksToDetections[i].type == Unassigned) {
-                tracksPositions.push_back(i);
-            }
-        }
-    }
-};
+#include "AssignmentsTable.h"
 
 // ---------------------------------------------------------------------------
 // Tracker manager tracks. Create, remove, update.
@@ -251,7 +38,7 @@ void Tracker::update(
     size_t N = tracks.size();        // tracks
     size_t M = detections.size();    // detects
 
-    Assignments assignments(N, M, distanceThreshold);
+    AssignmentsTable assignments(N, M, distanceThreshold);
 
     if (!tracks.empty()) {
         // Matrix of distance between N-th track and M-th detection.
@@ -300,7 +87,8 @@ void Tracker::update(
         }
 
         // =========================== Add occlusion handling ==========================
-//        occlusionHandler.update(tracks);
+        occlusionHandler.update(assignments, detections, tracks);
+
 
         assignments.solve(Cost, (int) N, (int) M);
 
@@ -323,10 +111,10 @@ void Tracker::update(
 //        }
 
         for (int i = 0; i < assignments.tracksToDetections.size(); ++i) {
-            if (assignments.tracksToDetections[i].type == Assignments::Unassigned) {
+            if (assignments.tracksToDetections[i].type == AssignmentsTable::Unassigned) {
                 tracks[i]->skipped_frames++;
                 if (tracks[i]->skipped_frames > maximum_allowed_skipped_frames) {
-                    assignments.tracksToDetections[i].type = Assignments::ToRemove;
+                    assignments.tracksToDetections[i].type = AssignmentsTable::ToRemove;
                 }
             }
         }
@@ -341,7 +129,7 @@ void Tracker::update(
 //        }
 //    }
     for (size_t i = 0; i < detections.size(); ++i) {
-        if (assignments.detectionsToTracks[i].type == Assignments::Unassigned) {
+        if (assignments.detectionsToTracks[i].type == AssignmentsTable::Unassigned) {
             tracks.push_back(std::make_unique<Track>(detections[i], dt, Accel_noise_mag, NextTrackID++));
         }
     }
@@ -355,17 +143,17 @@ void Tracker::update(
     for (int i = 0; i < assignments.detectionsToTracks.size(); ++i) {
         int detectionType = assignments.detectionsToTracks[i].type;
         if (detectionType ==
-            Assignments::Merged) { //add new track initialized by detection and merged tracks and delete all the merged tracks
+            AssignmentsTable::Merged) { //add new track initialized by detection and merged tracks and delete all the merged tracks
             std::vector<int> relatedTracks;
             for (int track : assignments.detectionsToTracks[i].additionalAssignments) {
                 relatedTracks.push_back(track);
                 tracksToRemove.insert(track);
             }
             tracks.push_back(std::make_unique<Track>(detections[i], dt, Accel_noise_mag,
-                                                     NextTrackID++, relatedTracks)); //TODO: add track constructor with n tracks
+                                                     NextTrackID++, relatedTracks));
         } else if (detectionType >= 0){
             if (assignments.tracksToDetections[detectionType].type ==
-                Assignments::Splitted) { //if the detection is assigned to track which is splitted then remove previous track and initialize new from it.
+                AssignmentsTable::Splitted) { //if the detection is assigned to track which is splitted then remove previous track and initialize new from it.
                 tracks.push_back(std::make_unique<Track>(detections[i], dt, Accel_noise_mag, NextTrackID++, detectionType));
                 tracksToRemove.insert(detectionType);
             }
@@ -374,7 +162,7 @@ void Tracker::update(
 
     // Mark tracks to remove with label ToRemove
     for (int i : tracksToRemove) {
-        assignments.tracksToDetections[i].type = Assignments::ToRemove;
+        assignments.tracksToDetections[i].type = AssignmentsTable::ToRemove;
     }
 
     // -----------------------------------
@@ -399,7 +187,7 @@ void Tracker::update(
             tracks[i]->skipped_frames = 0;
             tracks[i]->update(detections[detectionType], true, max_trace_length);
         } else if (detectionType ==
-                   Assignments::Unassigned)                  // if not continue using predictions of filter
+                   AssignmentsTable::Unassigned)                  // if not continue using predictions of filter
         {
             tracks[i]->update(Point_t(), false, max_trace_length);
         }
@@ -407,7 +195,7 @@ void Tracker::update(
 
     // Delete all the tracks with label ToRemove
     for (int i = 0; i < assignments.tracksToDetections.size(); i++) {
-        if (assignments.tracksToDetections[i].type == Assignments::ToRemove){
+        if (assignments.tracksToDetections[i].type == AssignmentsTable::ToRemove){
             tracks.erase(tracks.begin() + i);
             assignments.tracksToDetections.erase(assignments.tracksToDetections.begin() + i);
             i--;
